pyproject.toml
[tool.poetry]
name = "image-db-app"
version = "0.1.0"
description = "A short description of the project."
authors = ["Laurent <Your email address>"]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.21"
python-decouple = "^3.6"
SQLAlchemy = "^1.4.36"
Flask = "2.2.5"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate="^3.1.0"
json2table='^1.0.0'
Jinja2 = "^3.0.3"
MarkupSafe = "^2.1.0"
Werkzeug = "2.2.3"
click = "^8.0.4"
itsdangerous = "^2.1.2"
arrow = '^1.2.3'
flask-smorest = "^0.40.0"
marshmallow-sqlalchemy = "^0.28.1"
gunicorn = "^20.1.0"
Flask-FlatPages = "^0.8.1"
Pygments = "^2.13.0"
pyOpenSSL = "^22.1.0"
msal = "^1.20.0"
azure-ad-verify-token = "^0.3.0"
cachetools = "^5.2.0"
Bootstrap-Flask = "^2.2.0"
plotly = "^5.11.0"
psycopg2-binary = "^2.9.5"
SQLAlchemy-Utils = "^0.39.0"
Pillow = "^9.3.0"
numpy = "^1.25.2"
pandas = "^2.0.3"

[tool.poetry.group.dev.dependencies]
isort = "^5.2.2"
pytest-flask = "^1.2.0"
mock-file-tree = "^1.1.3"
pytest = "^7.2.2"
black = "^23.7.0"
pyright = "^1.1.322"
flake8 = "^6.1.0"
pylint = "^2.17.5"

[tool.pyright]
venv = "image-db-app"
exclude = [
    "**/__pycache__",
]
